[FEAT/#78] 옵션 선택 시 버튼 활성화

+++ presentation/src/main/java/co/orange/presentation/detail/OptionAdapter.kt
@@ -8,7 +8,7 @@ import co.orange.domain.entity.response.ProductOptionModel
 import kr.genti.presentation.databinding.ItemOptionBinding
 
 class OptionAdapter(
-    private val itemClick: (Long, Long) -> Unit,
+    private val itemClick: (Int, Long) -> Unit,
 ) : ListAdapter<ProductOptionModel, OptionViewHolder>(diffUtil) {
     override fun onCreateViewHolder(
         parent: ViewGroup,
